feat: 适配 Qwen3.5-2B 模型，新增 AUTO 量化模式并优化内存配置

+++ config.py
@@ -21,9 +21,10 @@ PROXY_URL = None
 # 模型路径 (本地)
 MODEL_PATH = "./models/Qwen3.5-2B"
 
-# 量化类型 ("INT4", "INT8", "FP16", "FP32")
+# 量化类型 ("INT4", "INT8", "FP16", "FP32", "AUTO")
 # 在 macOS/CPU 上，INT4/INT8 会被自动优化或回退
-QUANTIZATION = "INT8"
+# AUTO: 在 GPU 上使用 INT8，在 CPU 上使用 FP32（避免缓慢的动态量化）
+QUANTIZATION = "FP32"
 
 # 设备 ("cuda", "cpu", "mps")
 # 留空则自动检测
